refactor(utils): extract DOM helpers from utils.js
Move docById, docByClass, docByName, docBySelector, docByTagName,
closeWidgets, hideDOMLabel, and displayMsg out of utils.js into a new
js/utils/dom-helpers.js, following the same extraction pattern already
used for js/utils/utils-logic.js. utils.js now requires dom-helpers.js
and re-exports its members for CommonJS/Jest consumers; the RequireJS
shim (js/loader.js), the eager module list (js/activity.js), and the
non-RequireJS planet bootstrap (planet/index.html) are updated to load
the new file as a dependency of utils/utils. No other callers change,
since these functions are consumed as ambient globals elsewhere.
