+++ js/utils/utils.js
@@ -35,13 +35,17 @@ if (typeof module !== "undefined" && module.exports) {
         (typeof window !== "undefined" && window.UtilsLogic) ||
         (typeof require !== "undefined" ? require("./utils-logic") : {});
     var { resolveObject } = UtilsLogic;
+
+    var DomHelpers =
+        (typeof window !== "undefined" && window.DomHelpers) ||
+        (typeof require !== "undefined" ? require("./dom-helpers") : {});
 }
 
 /* exported
-   announceToScreenReader,canvasPixelRatio, changeImage, closeBlkWidgets, closeWidgets,
-   delayExecution, displayMsg, doBrowserCheck, docByClass, docByName,
-   docBySelector, docByTagName, doPublish, doStopVideoCam, doSVG,
-   doUseCamera, format, getTextWidth, hideDOMLabel, httpGet, httpPost, HttpRequest,
+   announceToScreenReader,canvasPixelRatio, changeImage, closeBlkWidgets,
+   delayExecution, doBrowserCheck,
+   doPublish, doStopVideoCam, doSVG,
+   doUseCamera, format, getTextWidth, httpGet, httpPost, HttpRequest,
    importMembers, isSVGEmpty, prepareMacroExports, preparePluginExports,
    processMacroData, processPluginData, processRawPluginData, windowHeight, windowWidth,
    fnBrowserDetect, waitForReadiness
@@ -439,50 +443,8 @@ function waitForReadiness(callback, options = {}) {
     requestAnimationFrame(check);
 }
 
-/**
- * Retrieves a collection of elements by class name.
- * @param {string} classname - The class name to search for.
- * @returns {HTMLCollectionOf<Element>} A collection of elements with the specified class name.
- */
-function docByClass(classname) {
-    return document.getElementsByClassName(classname);
-}
-
-/**
- * Retrieves a collection of elements by tag name.
- * @param {string} tag - The tag name to search for.
- * @returns {NodeList} A collection of elements with the specified tag name.
- */
-function docByTagName(tag) {
-    return document.getElementsByTagName(tag);
-}
-
-/**
- * Retrieves an element by its ID.
- * @param {string} id - The ID of the element to retrieve.
- * @returns {HTMLElement|null} The element with the specified ID, or null if not found.
- */
-function docById(id) {
-    return document.getElementById(id);
-}
-
-/**
- * Retrieves a collection of elements by name.
- * @param {string} name - The name attribute value to search for.
- * @returns {NodeListOf<Element>} A collection of elements with the specified name attribute.
- */
-function docByName(name) {
-    return document.getElementsByName(name);
-}
-
-/**
- * Retrieves the first element that matches a specified CSS selector.
- * @param {string} selector - A CSS selector string.
- * @returns {Element|null} The first element that matches the selector, or null if not found.
- */
-function docBySelector(selector) {
-    return document.querySelector(selector);
-}
+// docByClass(), docByTagName(), docById(), docByName(), docBySelector()
+// moved to js/utils/dom-helpers.js
 
 // last() and deepClone() moved to js/utils/utils-logic.js
 
@@ -1315,40 +1277,7 @@ function doStopVideoCam(cameraID, setCameraID) {
     CameraManager.reset();
 }
 
-/**
- * Hides certain DOM elements related to labels.
- */
-function hideDOMLabel() {
-    const textLabel = docById("textLabel");
-    if (textLabel !== null) {
-        textLabel.style.display = "none";
-    }
-
-    const numberLabel = docById("numberLabel");
-    if (numberLabel !== null) {
-        numberLabel.style.display = "none";
-    }
-
-    const piemenu = docById("wheelDiv");
-    if (piemenu !== null) {
-        piemenu.style.display = "none";
-    }
-}
-
-/**
- * Displays a message (currently unused).
- * @returns {undefined}
- */
-function displayMsg(/*blocks, text*/) {
-    /*
-    let msgContainer = blocks.msgText.parent;
-    msgContainer.visible = true;
-    blocks.msgText.text = text;
-    msgContainer.updateCache();
-    blocks.stage.setChildIndex(msgContainer, blocks.stage.getNumChildren() - 1);
-    */
-    return;
-}
+// hideDOMLabel() and displayMsg() moved to js/utils/dom-helpers.js
 
 // safeSVG() and toFixed2() moved to js/utils/utils-logic.js
 
@@ -1370,15 +1299,7 @@ let delayExecution = duration => {
     });
 };
 
-/**
- * Closes all widgets in the window.
- *
- * @returns {void}
- */
-function closeWidgets() {
-    const names = Object.keys(window.widgetWindows.openWindows);
-    names.forEach(name => window.widgetWindows.closeWindow(name));
-}
+// closeWidgets() moved to js/utils/dom-helpers.js
 
 /**
  * Closes a specific widget by its name.
@@ -1536,11 +1457,11 @@ let importMembers = (obj, className, modelArgs, viewArgs) => {
 if (typeof module !== "undefined" && module.exports) {
     module.exports = {
         ...UtilsLogic,
+        ...DomHelpers,
         extractProjectDataFromHTML,
         _,
         format,
         delayExecution,
-        closeWidgets,
         closeBlkWidgets,
         importMembers,
         changeImage,
@@ -1551,19 +1472,12 @@ if (typeof module !== "undefined" && module.exports) {
         httpGet,
         httpPost,
         HttpRequest,
-        docByClass,
-        docByTagName,
-        docById,
-        docByName,
-        docBySelector,
         getTextWidth,
         doSVG,
         isSVGEmpty,
         prepareMacroExports,
         processPluginData,
         processMacroData,
-        hideDOMLabel,
-        displayMsg,
         announceToScreenReader
     };
 }
