+++ planet/index.html
@@ -41,6 +41,7 @@
     <script type="text/javascript" src="js/LocalPlanet.js"></script>
 
     <script type="text/javascript" src="../js/utils/utils-logic.js"></script>
+    <script type="text/javascript" src="../js/utils/dom-helpers.js"></script>
     <script type="text/javascript" src="../js/utils/utils.js"></script>
     <script type="text/javascript" src="js/GlobalTag.js"></script>
     <script type="text/javascript" src="js/GlobalCard.js"></script>
